Delete our own copies of things, and use those from generator-common
